Added thoth.invokeEx
thoth.invokeEx will do the same as thoth.invoke except it takes a
callback that will decide if the execution continues or not. This
callback will recieve the return value of the las called callback and
also its index. Returning false will stop the execution of the
callbacks.

This allows to implement stopImmediatePropagation.